requirements.txt
discord.py==2.5.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
cachetools==5.5.0
pydantic==2.9.2
tenacity==8.5.0
openai==1.57.2
google-generativeai==0.8.3
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
